+++ plugins/actor.lua
@@ -17,5 +17,11 @@ Actor.fire_explosion = function()
 end
 
 
+Actor.damage = function(target, source, damage, x, y, color, crit_sfx)
+    if not crit_sfx then crit_sfx = false end
+    gm.damage_inflict(target, damage, 0.0, source, x, y, 0.0, 1.0, color or 16777215.0, crit_sfx)
+end
+
+
 Actor.heal = function()
 end